Effective to-do list controller using event delegation.
Event delegation was used -
instead of attaching events to individual imgs or divs, delegating it to their parent li helped in effective control on which lists to remove and which lists to keep.

A sample event bubbling control mechanism was also implemented using one of the element images to interfer with the delegated event to the parent list.